split out utils
